+++ src/types/_mimes.ts
