+++ src/utils/response.ts
@@ -1,13 +1,13 @@
 import type { Readable } from "node:stream";
 import type { Socket } from "node:net";
 import type { H3Event } from "../event";
-import type { HTTPHeaderName } from "../types";
-import type { ContentType } from "../types.mimes";
 import type {
+  HTTPHeaderName,
+  ContentType,
   HeaderValues,
   StatusCode,
   NodeHeaderValue,
-} from "../types.headers";
+} from "../types";
 import { MIMES } from "./consts";
 import { sanitizeStatusCode, sanitizeStatusMessage } from "./sanitize";
 import { splitCookiesString } from "./cookie";
